@@ -8,13 +8,14 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0|^8.1|^8.2",
+        "php": "^8.4",
         "deployer/deployer": "^7.0",
         "doctrine/dbal": "^3.1",
         "guzzlehttp/guzzle": "^7",
+        "irazasyed/telegram-bot-sdk": "^3.15",
         "laravel-notification-channels/telegram": "^6.0",
-        "laravel/framework": "^11.0",
-        "laravel/sanctum": "^4.0",
+        "laravel/framework": "^12.0",
+        "laravel/sanctum": "^4.1",
         "laravel/tinker": "^2",
         "laravel/ui": "^4.0",
         "league/html-to-markdown": "^5.0",
@@ -27,16 +28,16 @@
         "vlucas/phpdotenv": "^5.4"
     },
     "require-dev": {
-        "driftingly/rector-laravel": "^0.29.0",
+        "driftingly/rector-laravel": "^2.0",
         "fakerphp/faker": "^1.9.1",
+        "larastan/larastan": "^3.0",
         "laravel/pint": "^1.1",
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/larastan": "^2.0",
-        "pestphp/pest": "^2.0",
-        "pestphp/pest-plugin-laravel": "^2.0",
-        "phpunit/phpunit": "^10.1",
-        "rector/rector": "^0.18.0"
+        "pestphp/pest": "^3.8",
+        "pestphp/pest-plugin-laravel": "^3.2",
+        "phpunit/phpunit": "^11.0",
+        "rector/rector": "^2.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -44,7 +45,8 @@
         "sort-packages": true,
         "allow-plugins": {
             "pestphp/pest-plugin": true,
-            "php-http/discovery": false
+            "php-http/discovery": false,
+            "thecodingmachine/discovery": true
         }
     },
     "extra": {
@@ -81,7 +83,7 @@
             "@php artisan key:generate --ansi"
         ],
         "test": [
-            "vendor/bin/phpunit"
+            "vendor/bin/pest"
         ]
     }
 }
