@@ -29,12 +29,11 @@ class FilesController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\StoreFile  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(\App\Http\Requests\StoreFile $request)
     {
-        $file = self::storeFile($request);
+        $file = $this->storeFile($request);
 
         return response()->json($file, 201);
     }
@@ -75,22 +74,22 @@ class FilesController extends Controller
         $file->delete();
 
         // Delete cache if necessary
-        if (property_exists($file, 'attachable') && $file->attachable !== null && get_class($file->attachable) == 'App\Models\Event') {
+        if ($file->attachable_type === Event::class) {
             Cache::forget(Event::$CACHE_KEY_CURRENT_EVENTS);
             Cache::forget(Event::$CACHE_KEY_GROUPED_EVENTS);
         }
 
-        if (property_exists($file, 'attachable') && $file->attachable !== null && get_class($file->attachable) == 'App\Models\News') {
+        if ($file->attachable_type === News::class) {
             Cache::forget(News::$CACHE_KEY_TOP_NEWS);
         }
 
         return response()->json(null, 204);
     }
 
-    public static function storeFile(Request $request)
+    public function storeFile(Request $request)
     {
         if (! $request->hasFile('file')) {
-            return;
+            return null;
         }
 
         // Collect file information
