@@ -15,7 +15,9 @@ class Authenticate extends Middleware
     protected function redirectTo($request)
     {
         if (! $request->expectsJson()) {
-            return response()->json('Accept: application/json header missing', 400);
+            return route('login');
         }
+
+        return null;
     }
 }
