@@ -22,7 +22,7 @@ class StoreFile extends FormRequest
     public function rules()
     {
         return [
-            'file' => 'file',
+            'file' => 'required|file',
         ];
     }
 }
