@@ -2,9 +2,10 @@
 
 namespace App\Services;
 
-use Analytics;
 use Carbon\Carbon;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Cache;
+use Spatie\Analytics\Facades\Analytics;
 use Spatie\Analytics\Period;
 
 class AnalyticsService
@@ -18,7 +19,7 @@ class AnalyticsService
 
     public function retrieveAnalyticsData($year, $month)
     {
-        $cache_key = "analytics_${year}_${month}";
+        $cache_key = "analytics_{$year}_{$month}";
 
         return Cache::remember($cache_key, $this->cache_ttl, function () use ($year, $month) {
             // Data for requested month
@@ -27,10 +28,10 @@ class AnalyticsService
             $period = Period::create($startDate, $endDate);
 
             $requestedAnalyticsData = $this->getMetrics($period);
-            $mostVisitedPages = Analytics::fetchMostVisitedPages($period, 5)->toArray();
-            $topBrowsers = Analytics::fetchTopBrowsers($period, 5)->toArray();
-            $topReferrers = Analytics::fetchTopReferrers($period, 5)->toArray();
-            $userTypes = Analytics::fetchUserTypes($period)->toArray();
+            $mostVisitedPages = Analytics::fetchMostVisitedPages($period, 5);
+            $topBrowsers = Analytics::fetchTopBrowsers($period, 5);
+            $topReferrers = Analytics::fetchTopReferrers($period, 5);
+            $userTypes = Analytics::fetchUserTypes($period);
 
             // Data for previous month
             $startDate = $startDate->subMonth();
@@ -42,33 +43,34 @@ class AnalyticsService
             return [
                 'previousMonth' => $previousAnalyticsData,
                 'requestedMonth' => $requestedAnalyticsData,
-                'mostVisitedPages' => $mostVisitedPages,
-                'topBrowsers' => $topBrowsers,
-                'topReferrers' => $topReferrers,
-                'userTypes' => $userTypes,
+                'mostVisitedPages' => $this->toArray($mostVisitedPages),
+                'topBrowsers' => $this->toArray($topBrowsers),
+                'topReferrers' => $this->toArray($topReferrers),
+                'userTypes' => $this->toArray($userTypes),
             ];
         });
     }
 
     protected function getMetrics(Period $period)
     {
-        $totals = Analytics::performQuery(
+        $data = Analytics::get(
             $period,
-            implode(',', [
-                'ga:visitors',
-                'ga:newUsers',
-                'ga:sessions',
-                'ga:bounceRate',
-                'ga:avgSessionDuration',
-                'ga:organicSearches',
-            ])
-        )->totalsForAllResults;
+            ['activeUsers', 'sessions', 'bounceRate', 'averageSessionDuration', 'screenPageViews'],
+            ['pageTitle']
+        );
+
+        $data = $data instanceof Collection ? $data->first() : $data[0];
 
         return [
-            'visitors' => (int) $totals['ga:visitors'],
-            'sessions' => (int) $totals['ga:sessions'],
-            'bounceRate' => (float) $totals['ga:bounceRate'] / 100,
-            'avgSessionDurationInSeconds' => (int) $totals['ga:avgSessionDuration'],
+            'visitors' => (int) $data['activeUsers'],
+            'sessions' => (int) $data['sessions'],
+            'bounceRate' => (float) $data['bounceRate'] / 100,
+            'avgSessionDurationInSeconds' => (int) $data['averageSessionDuration'],
         ];
+    }
+
+    protected function toArray($data)
+    {
+        return $data instanceof Collection ? $data->toArray() : $data;
     }
 }
