@@ -5,7 +5,7 @@ return [
     /*
      * The view id of which you want to display data.
      */
-    'view_id' => env('ANALYTICS_VIEW_ID', '12345678'),
+    'property_id' => env('ANALYTICS_VIEW_ID', '12345678'),
 
     /*
      * Path to the client secret json file. Take a look at the README of this package
