@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 
 class AddTemplateDefaultValue extends Migration
@@ -13,10 +14,13 @@ class AddTemplateDefaultValue extends Migration
     public function up()
     {
         Schema::table('tbl_site', function ($table) {
-            $table->text('template')->nullable()->default('template.php')->change();
+            $table->text('template')->nullable()->change();
             $table->text('template_name')->nullable()->change();
             $table->text('description')->nullable()->change();
         });
+
+        // Set default value after column modification
+        DB::table('tbl_site')->whereNull('template')->update(['template' => 'template.php']);
     }
 
     /**
@@ -27,8 +31,8 @@ class AddTemplateDefaultValue extends Migration
     public function down()
     {
         Schema::table('tbl_site', function ($table) {
-            $table->text('template')->nullable(false)->default('')->change();
             $table->text('template')->nullable(false)->change();
+            $table->text('template_name')->nullable(false)->change();
             $table->text('description')->nullable(false)->change();
         });
     }
