@@ -2,7 +2,6 @@
 
 use App\Models\News;
 use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
 
@@ -11,8 +10,6 @@ use function Pest\Laravel\deleteJson;
 use function Pest\Laravel\getJson;
 use function Pest\Laravel\postJson;
 use function Pest\Laravel\putJson;
-
-uses(RefreshDatabase::class);
 
 test('it returns a list of news', function () {
     $user = User::factory()->create();
@@ -95,6 +92,38 @@ test('it deletes a news article', function () {
 
     $response->assertStatus(204);
     $this->assertDatabaseMissing('tbl_news', ['id' => $news->ID]);
+});
+
+test('it deletes a news article with attachments', function () {
+    $user = User::factory()->admin()->create();
+    actingAs($user, 'sanctum');
+
+    Storage::fake('local');
+    $news = News::factory()->create();
+    $file = UploadedFile::fake()->create('test.jpg');
+
+    // Mock the FilesController
+    $fileControllerMock = $this->mock(\App\Http\Controllers\Api\FilesController::class)->makePartial();
+    $fileModel = new \App\Models\File(['file' => 'test.jpg', 'title' => 'test.jpg']);
+    $fileControllerMock->shouldReceive('storeFile')->andReturn($fileModel);
+
+    // Attach the file
+    $request = new \Illuminate\Http\Request;
+    $request->files->set('file', $file);
+    $fileRecord = $fileControllerMock->storeFile($request);
+    $news->attachments()->save($fileRecord);
+
+    Storage::disk('local')->put('test.jpg', 'test');
+    Storage::disk('local')->assertExists($fileRecord->file);
+
+    // Delete the news
+    $response = $this->deleteJson("/api/news/{$news->ID}");
+    $response->assertStatus(204);
+
+    // Assertions
+    $this->assertDatabaseMissing('tbl_news', ['ID' => $news->ID]);
+    $this->assertDatabaseMissing('tbl_downloads', ['ID' => $fileRecord->ID]);
+    Storage::disk('local')->assertMissing($fileRecord->file);
 });
 
 test('it attaches a file to a news article', function () {
