@@ -3,43 +3,53 @@
 namespace Tests\Feature;
 
 use App\Models\User;
-use Tests\TestCase;
 
-class LoginTest extends TestCase
-{
-    private $user;
+use function Pest\Laravel\postJson;
 
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->user = User::factory()->create();
-
-        // Session needs to be started before login works.
-        $kernel = app('Illuminate\Contracts\Http\Kernel');
-        $kernel->pushMiddleware('Illuminate\Session\Middleware\StartSession');
-    }
-
-    public function testRequiresEmailAndPassword()
-    {
-        $this->postJson('api/login')
-            ->assertStatus(422)
-            ->assertJson([
-                'errors' => [
-                    'email' => ['validation.required'],
-                    'password' => ['validation.required'],
-                ],
-            ]);
-    }
-
-    public function testUserLogsInSuccessfully()
-    {
-        $response = $this->post('api/login', [
-            'email' => $this->user->email,
-            'password' => 'password',
+test('requires email and password', function () {
+    postJson('api/login')
+        ->assertStatus(422)
+        ->assertJson([
+            'errors' => [
+                'email' => ['validation.required'],
+                'password' => ['validation.required'],
+            ],
         ]);
+});
 
-        $response->assertStatus(200);
-        $this->assertAuthenticatedAs($this->user);
-    }
-}
+test('user logs in successfully', function () {
+    $user = User::factory()->create([
+        'password' => bcrypt($password = 'i-love-laravel'),
+    ]);
+
+    $response = postJson('api/login', [
+        'email' => $user->email,
+        'password' => $password,
+    ]);
+
+    $response->assertStatus(200);
+    $this->assertAuthenticatedAs($user);
+    $response->assertJsonFragment(['email' => $user->email]);
+});
+
+test('user can log out', function () {
+    $user = User::factory()->create();
+    $this->actingAs($user);
+
+    $response = postJson('api/logout');
+
+    $response->assertStatus(200);
+    $this->assertGuest();
+});
+
+test('it returns an error with invalid credentials', function () {
+    $user = User::factory()->create();
+
+    $response = postJson('api/login', [
+        'email' => $user->email,
+        'password' => 'wrong-password',
+    ]);
+
+    $response->assertStatus(401);
+    $response->assertJsonFragment(['email' => ['auth.failed']]);
+});
