@@ -11,10 +11,7 @@
 |
 */
 
-// Uses the given trait in the current file
-uses(Illuminate\Foundation\Testing\RefreshDatabase::class);
-uses(Tests\TestCase::class)->in('Unit');
-uses(Tests\TestCase::class)->in('Feature');
+uses(Tests\TestCase::class)->in('Feature', 'Unit');
 
 /*
 |--------------------------------------------------------------------------
